+++ database/src/models/game/bgm.rs
@@ -5,7 +5,7 @@ use sqlx::FromRow;
 pub struct UserBgm {
     pub player_id: i64,
     pub bgm_id: i32,
-    pub unlock_time: i32,
+    pub unlock_time: i64,
     pub is_favorite: bool,
     pub is_read: bool,
 }
@@ -14,7 +14,7 @@ impl From<UserBgm> for sonettobuf::BgmInfo {
     fn from(bgm: UserBgm) -> Self {
         sonettobuf::BgmInfo {
             bgm_id: Some(bgm.bgm_id),
-            unlock_time: Some(bgm.unlock_time),
+            unlock_time: Some(bgm.unlock_time as i32),
             favorite: Some(bgm.is_favorite),
             is_read: Some(bgm.is_read),
         }
